+++ Runtime/Scripts/InworldClient.cs
@@ -1,5 +1,5 @@
 ﻿/*************************************************************************************************
- * Copyright 2022-2024 Theai, Inc. dba Inworld AI
+ * Copyright 2022-2025 Theai, Inc. dba Inworld AI
  *
  * Use of this source code is governed by the Inworld.ai Software Development Kit License Agreement
  * that can be found in the LICENSE.md file or at https://www.inworld.ai/sdk-license
@@ -882,6 +882,7 @@ namespace Inworld
         {
             if (string.IsNullOrEmpty(base64))
                 return false;
+            
             DataChunk dataChunk = new DataChunk
             {
                 type = DataType.AUDIO,