+++ Runtime/Scripts/UI/RecordButton.cs
@@ -1,5 +1,5 @@
 /*************************************************************************************************
- * Copyright 2022-2024 Theai, Inc. dba Inworld AI
+ * Copyright 2022-2025 Theai, Inc. dba Inworld AI
  *
  * Use of this source code is governed by the Inworld.ai Software Development Kit License Agreement
  * that can be found in the LICENSE.md file or at https://www.inworld.ai/sdk-license
@@ -15,11 +15,11 @@ namespace Inworld.UI
         {
             if (InworldController.Client.Status == InworldConnectionStatus.Connected)
                 InworldController.Instance.CancelResponses();
-            InworldController.Audio.IsRecording = true;
+            InworldController.Audio.IsPlayerSpeaking = true;
         }
         public void OnPointerUp(PointerEventData eventData)
         {
-            InworldController.Audio.IsRecording = false;
+            InworldController.Audio.IsPlayerSpeaking = false;
         }
     }
 }